pyproject.toml
[tool.poetry]
name = "ecodev-core"
version = "0.0.1"
description = "Low level sqlmodel/fastapi/pydantic building blocks"
authors = ["Thomas Epelbaum <tomepel@gmail.com>",
					 "Olivier Gabriel <olivier.gabriel.geom@gmail.com>",
					 "Amaury Salles <amaury.salles@gmail.com>",
					 "Yoann Diep <yoann.diep@hotmail.fr>",
					 "Dorian Kodelja <dorian.kodelja@gmail.com>",

]
license = "MIT"
readme = "README.md"
classifiers = [
    "Intended Audience :: Information Technology",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Intended Audience :: System Administrators",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python",
    "Topic :: Internet",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development",
    "Typing :: Typed",
    "Development Status :: 2 - Pre-Alpha",
    "Environment :: Web Environment",
    "Framework :: FastAPI",
    "Framework :: Pydantic",
    "Framework :: AsyncIO",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Database",
    "Topic :: Database :: Database Engines/Servers"
]
[tool.poetry.dependencies]
python = "^3.11"
fastapi = "~0"
pandas = "~2"
sqlmodel = "~0"
numpy = "~1"
pydantic = {version = "~1", extras = ["python-dotenv"]}
python-jose = {version = "~3", extras = ["cryptography"]}
passlib = {version = "~1", extras = ["bcyrypt"]}
sqladmin = "0.15.2"
httpx = "~0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
